Fix invalid slice parameter when rendering branch tree
When the branch tree is rendered each branch is given a "context" of
commits that must be displayed. The master branch is fixed to 3 and all
other branches are fully rendered as they contain only commits that are
only in them.

The rendering function is called recursively with initial context of 3 and
each recursive call is done with a context of -1. There lies the problem.
Slice [:-1] gives everything but the last item which causes branches with
a single commit to appear empty.

Slice [:None] on the other hand does exactly what is needed.