Detect & render "alias" branches.
Branches that are merged together are now detected and rendered.
Such branches will be called "alias" branches. Each branch will now
have a name and aliases. The name will be used to identify that branch.

The name of the merged branch that happens to be detected first is
selected as a name of all the "alias" branhces (depends on the order
that 'git log' will list each branch). If one of the aliases is 'master'
then the name is changed to 'master' even if it is not the first in order.

The remaining names are ordered by their respective length. If the line
gets too long, the longest names will be grouped together and displayed
as a number of hidden names (e.g. [ +4 more ]).

Remote branch names are completely hidden unless no local aliases exist.